Jellyfin Playlist Generator
A standalone html/javascript Jellyfin playlist generator. Just download the files into a folder, run index.html and enter your server and API key. It will fetch a list of users and then provide the option to create a filtered list of movies and/or tv shows. Add and remove titles as you please and add to the master list. Once you are satisfied with the contents of the master list, give your playlist a name and click save playlist to Jellyfin for currently selected user.